variant file, module and component names changed
