+++ projects/storefrontlib/src/cms-components/cms-lib.module.ts
@@ -50,7 +50,7 @@ import {
 } from './product/index';
 import { ProductImagesModule } from './product/product-images/product-images.module';
 import { ProductSummaryModule } from './product/product-summary/product-summary.module';
-import { ProductVariantSelectorModule } from './product/product-variant-selector/product-variant-selector.module';
+import { ProductVariantsModule } from './product/product-variants/product-variants.module';
 import { StoreFinderModule } from './storefinder/store-finder.module';
 import { UserComponentModule } from './user/user.module';
 import { WishListModule } from './wish-list/wish-list.module';
@@ -93,7 +93,7 @@ import { WishListModule } from './wish-list/wish-list.module';
     StoreFinderModule,
     ProductImagesModule,
     ProductSummaryModule,
-    ProductVariantSelectorModule,
+    ProductVariantsModule,
     ProductIntroModule,
     CheckoutComponentModule,
     ForgotPasswordModule,
