+++ projects/storefrontlib/src/cms-components/product/product-list/container/product-scroll/product-scroll.component.spec.ts
@@ -148,7 +148,7 @@ export class MockProductListComponentService {
 }
 
 @Component({
-  selector: 'cx-style-icons',
+  selector: 'cx-variant-style-icons',
   template: 'test',
 })
 export class MockStyleIconsComponent {
