+++ projects/storefrontlib/src/cms-components/product/product-list/product-grid-item/product-grid-item.component.html
@@ -32,8 +32,10 @@
 </div>
 
 <ng-container *cxFeatureLevel="'1.5'">
-  <div class="cx-style-icons" *ngIf="product.variantOptions">
-    <cx-style-icons [variants]="product.variantOptions"></cx-style-icons>
+  <div class="cx-variant-style-icons" *ngIf="product.variantOptions">
+    <cx-variant-style-icons
+      [variants]="product.variantOptions"
+    ></cx-variant-style-icons>
   </div>
 </ng-container>
