+++ projects/storefrontlib/src/cms-components/product/product-list/product-grid-item/product-grid-item.component.spec.ts
@@ -55,7 +55,7 @@ class MockUrlPipe implements PipeTransform {
 }
 
 @Component({
-  selector: 'cx-style-icons',
+  selector: 'cx-variant-style-icons',
   template: 'test',
 })
 export class MockStyleIconsComponent {
