+++ projects/storefrontlib/src/cms-components/product/product-list/product-list-item/product-list-item.component.html
@@ -45,10 +45,10 @@
     </div>
     <ng-container *cxFeatureLevel="'1.5'">
       <div class="row" *ngIf="product.variantOptions">
-        <cx-style-icons
+        <cx-variant-style-icons
           class="list-style-icons"
           [variants]="product.variantOptions"
-        ></cx-style-icons>
+        ></cx-variant-style-icons>
       </div>
     </ng-container>
   </div>
