+++ projects/storefrontlib/src/cms-components/product/product-list/product-list-item/product-list-item.component.spec.ts
@@ -54,7 +54,7 @@ class MockUrlPipe implements PipeTransform {
 }
 
 @Component({
-  selector: 'cx-style-icons',
+  selector: 'cx-variant-style-icons',
   template: 'test',
 })
 export class MockStyleIconsComponent {
