+++ projects/storefrontlib/src/cms-components/product/product-variants/product-variants.component.html
@@ -1,18 +1,18 @@
 <ng-container *ngIf="product$ | async as product">
   <div class="variant-section" *ngIf="product.baseOptions?.length">
-    <cx-style-selector
+    <cx-variant-style-selector
       *ngIf="variants[variantType.STYLE]"
       [variants]="variants[variantType.STYLE]"
-    ></cx-style-selector>
-    <cx-size-selector
+    ></cx-variant-style-selector>
+    <cx-variant-size-selector
       *ngIf="variants[variantType.SIZE]"
       [product]="product"
       [variants]="variants[variantType.SIZE]"
-    ></cx-size-selector>
-    <cx-color-selector
+    ></cx-variant-size-selector>
+    <cx-variant-color-selector
       *ngIf="variants[variantType.COLOR]"
       [product]="product"
       [variants]="variants[variantType.COLOR]"
-    ></cx-color-selector>
+    ></cx-variant-color-selector>
   </div>
 </ng-container>
