+++ projects/storefrontlib/src/cms-components/product/product-variants/variant-color-selector/variant-color-selector.component.html
