+++ projects/storefrontlib/src/cms-components/product/product-variants/variant-size-selector/variant-size-selector.component.html
