+++ projects/storefrontlib/src/cms-components/product/product-variants/variant-size-selector/variant-size-selector.component.ts
@@ -8,8 +8,8 @@ import {
 } from '@spartacus/core';
 
 @Component({
-  selector: 'cx-size-selector',
-  templateUrl: './size-selector.component.html',
+  selector: 'cx-variant-size-selector',
+  templateUrl: './variant-size-selector.component.html',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class VariantSizeSelectorComponent {
