+++ projects/storefrontlib/src/cms-components/product/product-variants/variant-size-selector/variant-size-selector.module.ts
@@ -1,13 +1,13 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { UrlModule, I18nModule } from '@spartacus/core';
-import { VariantStyleSelectorComponent } from './style-selector.component';
+import { VariantSizeSelectorComponent } from './variant-size-selector.component';
 import { RouterModule } from '@angular/router';
 
 @NgModule({
   imports: [CommonModule, RouterModule, UrlModule, I18nModule],
-  declarations: [VariantStyleSelectorComponent],
-  entryComponents: [VariantStyleSelectorComponent],
-  exports: [VariantStyleSelectorComponent],
+  declarations: [VariantSizeSelectorComponent],
+  entryComponents: [VariantSizeSelectorComponent],
+  exports: [VariantSizeSelectorComponent],
 })
-export class VariantStyleSelectorModule {}
+export class VariantSizeSelectorModule {}
