+++ projects/storefrontlib/src/cms-components/product/product-variants/variant-style-icons/variant-style-icons.component.scss
