+++ projects/storefrontlib/src/cms-components/product/product-variants/variant-style-icons/variant-style-icons.module.ts
@@ -1,13 +1,13 @@
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { UrlModule, I18nModule } from '@spartacus/core';
-import { VariantColorSelectorComponent } from './color-selector.component';
+import { VariantStyleIconsComponent } from './variant-style-icons.component';
 import { RouterModule } from '@angular/router';
 
 @NgModule({
   imports: [CommonModule, RouterModule, UrlModule, I18nModule],
-  declarations: [VariantColorSelectorComponent],
-  entryComponents: [VariantColorSelectorComponent],
-  exports: [VariantColorSelectorComponent],
+  declarations: [VariantStyleIconsComponent],
+  entryComponents: [VariantStyleIconsComponent],
+  exports: [VariantStyleIconsComponent],
 })
-export class VariantColorSelectorModule {}
+export class VariantStyleIconsModule {}
