+++ projects/storefrontlib/src/cms-components/product/product-variants/variant-style-selector/variant-style-selector.component.html
