+++ projects/storefrontlib/src/cms-components/product/product-variants/variant-style-selector/variant-style-selector.component.spec.ts
@@ -7,7 +7,7 @@ import {
   BaseOption,
   VariantType,
 } from '@spartacus/core';
-import { VariantStyleSelectorComponent } from './style-selector.component';
+import { VariantStyleSelectorComponent } from './variant-style-selector.component';
 import { Pipe, PipeTransform } from '@angular/core';
 
 const mockOccBackendUrl = 'abc';
